1.完成了AbstractControllerV1 2.完成了AbstractService

+++ src/main/java/com/ming/site/service/ProductImageService.java
@@ -3,5 +3,5 @@ package com.ming.site.service;
 import com.ming.site.model.ProductImage;
 
 public interface ProductImageService
-        extends CrudService<ProductImage, Integer> {
+        extends CrudService<ProductImage, Long> {
 }
